// web/nestspend-web/src/app/core/api/models/category-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Category response data
 */
export interface CategoryResponse {

  /**
   * Color code for the category
   */
  color?: string;

  /**
   * Timestamp when the category was created
   */
  createdAt?: string;

  /**
   * Icon identifier for the category
   */
  icon?: string;

  /**
   * Unique identifier of the category
   */
  id?: string;

  /**
   * Name of the category
   */
  name?: string;
}