// web/nestspend-web/src/app/core/api/models/category-expense-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Category expense breakdown for monthly dashboard
 */
export interface CategoryExpenseResponse {

  /**
   * Total expense amount in cents for this category
   */
  amountCents?: number;

  /**
   * UUID of the category
   */
  categoryId?: string;

  /**
   * Name of the category
   */
  categoryName?: string;
}